% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_starts.R
\name{addIndicators}
\alias{addIndicators}
\title{Add Indicators}
\usage{
addIndicators(df, var, indicators, labelType = "numbers", sd = 1)
}
\arguments{
\item{df}{A dataframe, usually generated by \code{gen_starts()}. Variables should
be names X1 to Xw and Y1 to Yw, where w is the last wave.}

\item{var}{Numeric value representing the residual variance for the
indicators.}

\item{indicators}{Numeric value representing the number of indicators to add.}

\item{labelType}{Either "numbers" (the default) or "letters"}

\item{sd}{Numeric value indicating the standard deviation of the residual
variance that is added.}
}
\value{
Dataframe with the original variables and the new indicators.
}
\description{
\code{addIndicators()} takes data generated using \code{gen_starts()} and adds
indicators for each measure at each wave. The amount of unique variance
in the indicators can be specified (though this has to be the same for all
indicators.
}
