% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildMplus.R
\name{buildConstraints}
\alias{buildConstraints}
\title{Build Constraints}
\usage{
buildConstraints(
  waves,
  XVar = TRUE,
  YVar = TRUE,
  trait = TRUE,
  AR = TRUE,
  state = TRUE,
  crossLag = TRUE,
  stateCor = TRUE,
  stationarity = TRUE,
  constrainCor = stationarity,
  limits = TRUE
)
}
\arguments{
\item{waves}{Numeric value indicating total number of waves.}

\item{XVar}{Logical. Include X variable.}

\item{YVar}{logical. Include Y variable.}

\item{trait}{Logical value indicating whether to include a stable trait.
Defaults to TRUE.}

\item{AR}{logical value indicating whether to include autoregressive trait.
Defaults to TRUE.}

\item{state}{logical value indicating whether to include state. Defaults to
TRUE.}

\item{crossLag}{logical value indicating whether to include cross-lagged
paths. Defaults to TRUE.}

\item{stateCor}{logical value indicating whether to include state
correlations. Defaults to TRUE.}

\item{stationarity}{logical value indicating whether to impose stationarity.
Defaults to TRUE.}
}
\value{
Character vector representing the Mplus code for the constraints
statement.
}
\description{
Builds constraint statement for Mplus input file based on specified options.
}
