% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildDpm.R
\name{buildConstraints_d}
\alias{buildConstraints_d}
\title{Build Constraints for DPM}
\usage{
buildConstraints_d(
  waves,
  XVar = TRUE,
  YVar = TRUE,
  crossLag = TRUE,
  arCor = TRUE,
  limits = TRUE
)
}
\arguments{
\item{waves}{Numeric value indicating total number of waves.}

\item{XVar}{Logical. Include X variable.}

\item{YVar}{logical. Include Y variable.}

\item{crossLag}{logical value indicating whether to include cross-lagged
paths. Defaults to TRUE.}

\item{arCor}{Logical value indicating whether to include within-wave
correlations for AR traits.}

\item{limits}{Logical value indicating whether to constrain variances and
correlations to valid values. Defaults to TRUE.}
}
\value{
Character vector representing the Mplus code for the constraints
statement.
}
\description{
Builds constraint statement for Mplus input file based on specified options.
}
