% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildDpm.R
\name{buildDpm}
\alias{buildDpm}
\title{Builds Mplus Code for Dynamic Panel Model}
\usage{
buildDpm(
  waves,
  XVar = TRUE,
  YVar = TRUE,
  xIndicators = 1,
  yIndicators = 1,
  crossLag = TRUE,
  arCor = TRUE,
  constrainCors = TRUE
)
}
\arguments{
\item{waves}{Numeric value indicating total number of waves.}

\item{XVar}{Logical. Include X variable.}

\item{YVar}{logical. Include Y variable.}

\item{xIndicators}{Numeric value indicatoring number of indicators for X.
Defaults to 1.}

\item{yIndicators}{Numeric value indicatoring number of indicators for Y.
Defaults to 1.}

\item{crossLag}{logical value indicating whether to include cross-lagged
paths. Defaults to TRUE.}

\item{arCor}{logical value indicating whether to include autoregressive trait
correlations. Defaults to TRUE.}

\item{constrainCors}{logical value indicating whether to constrain
correlations between same indicator at different waves (when there are more
than one indicator.}
}
\value{
Character vector representing the Mplus code for the model statement.
}
\description{
\code{buildDpm()} produces code for variations of the Dynamic Panel Model.
Univariate and bivariate versions can be specified. Cross-lagged paths are
included by default, though they can be removed. Correlations between state
and autoregressive components can also be included.
}
