% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildModel.R
\name{buildLavaan}
\alias{buildLavaan}
\title{Build Lavaan Code}
\usage{
buildLavaan(
  waves,
  xWaves = NULL,
  yWaves = NULL,
  XVar = TRUE,
  YVar = TRUE,
  stationarity = TRUE,
  trait = TRUE,
  AR = TRUE,
  state = TRUE,
  crossLag = TRUE,
  stateCor = FALSE,
  limits = TRUE
)
}
\arguments{
\item{waves}{Numeric value indicating total number of waves.}

\item{xWaves}{Vector of actual waves for X (omit if same as waves).}

\item{yWaves}{Vector of actual waves for Y (omit if same as waves).}

\item{XVar}{Logical. Include X variable.}

\item{YVar}{logical. Include Y variable.}

\item{stationarity}{logical value indicating whether to impose stationarity.
Defaults to TRUE.}

\item{trait}{Logical value indicating whether to include a stable trait.
Defaults to TRUE.}

\item{AR}{logical value indicating whether to include autoregressive trait.
Defaults to TRUE.}

\item{state}{logical value indicating whether to include state. Defaults to
TRUE.}

\item{crossLag}{logical value indicating whether to include cross-lagged
paths. Defaults to TRUE.}

\item{stateCor}{logical value indicating whether to include state
correlations. Defaults to TRUE.}

\item{limits}{Logical value indicating whether to constrain variances and
correlations to possible values.}
}
\value{
Character vector representing the Mplus code for the model statement.
}
\description{
\code{buildLavaan()} produces lavaan code for variations of the general Stable Trait,
Autoregressive Trait, State Model. Univariate and bivariate versions can
be specified. Reduced versions of the model can be specified by setting
variance components for trait, AR, or state to be \code{FALSE}.
}
