% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildMplus.R
\name{buildMplus}
\alias{buildMplus}
\title{Builds Mplus Code}
\usage{
buildMplus(
  waves,
  XVar = TRUE,
  YVar = TRUE,
  xWaves = NULL,
  yWaves = NULL,
  xIndicators = 1,
  yIndicators = 1,
  trait = TRUE,
  AR = TRUE,
  state = TRUE,
  crossLag = TRUE,
  stateCor = FALSE,
  stationarity = TRUE
)
}
\arguments{
\item{waves}{Numeric value indicating total number of waves.}

\item{XVar}{Logical. Include X variable.}

\item{YVar}{logical. Include Y variable.}

\item{xWaves}{Vector of actual waves for X (omit if same as waves).}

\item{yWaves}{Vector of actual waves for Y (omit if same as waves).}

\item{xIndicators}{Numeric value indicatoring number of indicators for X.
Defaults to 1.}

\item{yIndicators}{Numeric value indicatoring number of indicators for Y.
Defaults to 1.}

\item{trait}{Logical value indicating whether to include a stable trait.
Defaults to TRUE.}

\item{AR}{logical value indicating whether to include autoregressive trait.
Defaults to TRUE.}

\item{state}{logical value indicating whether to include state. Defaults to
TRUE.}

\item{crossLag}{logical value indicating whether to include cross-lagged
paths. Defaults to TRUE.}

\item{stateCor}{logical value indicating whether to include state
correlations. Defaults to TRUE.}

\item{stationarity}{logical value indicating whether to impose stationarity.
Defaults to TRUE.}
}
\value{
Character vector representing the Mplus code for the model statement.
}
\description{
\code{buildMplus()} produces code for variations of the general Stable Trait,
Autoregressive Trait, State Model. Univariate and bivariate versions can
be specified. Reduced versions of the model can be specified by setting
variance components for trait, AR, or state to be \code{FALSE}.
}
