% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareModels.R
\name{compareModels}
\alias{compareModels}
\title{Run and Compare Models Estimated Using panelCodeR}
\usage{
compareModels(
  data,
  program = "mplus",
  models = list("starts", "riclpm", "arts", "sts", "clpm"),
  title = NULL,
  ...
)
}
\arguments{
\item{data}{Dataframe with multiwave data.}

\item{program}{Name of the program. Defaults to mplus as it is much quicker.}

\item{models}{List of models to compare. Defaults to list("starts", "riclpm",
"arts", "sts", "clpm"). Can also include "dpm_c", "dpm_p", "gclm", "alt",
"lgm_sm", or "lgcm" (see panelCodeR documentation).}

\item{title}{Optional character string to be attended to title of model.}

\item{...}{Options to pass to panelcoder function.}
}
\description{
\code{compareModels()} produces code for, runs, and compares various models for
analyzing panel data. It produces a table of basic results.
}
