% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewrite.R
\name{.buildModel}
\alias{.buildModel}
\title{Build Model}
\usage{
.buildModel(
  data,
  panelModel = "starts",
  crossLag = TRUE,
  stateCors = FALSE,
  residCors = FALSE,
  limits = TRUE,
  stationarity = TRUE
)
}
\description{
Builds parameter table representing model
}
