% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildParTable.R
\name{.buildTable}
\alias{.buildTable}
\title{Build Table}
\usage{
.buildTable(
  info,
  ar = TRUE,
  trait = TRUE,
  stability = TRUE,
  cl = TRUE,
  state = TRUE,
  traitCors = TRUE,
  arCors = TRUE,
  stateCors = TRUE,
  residCors = TRUE
)
}
\description{
Builds parameter table from model specification
}
