% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gen_starts}
\alias{gen_starts}
\title{Generate STARTS Data}
\usage{
gen_starts(
  n = 500,
  nwaves = 10,
  ri_x = 1,
  ri_y = 1,
  cor_i = 0.5,
  x = 1,
  y = 1,
  stab_x = 0.5,
  stab_y = 0.5,
  yx = 0.4,
  xy = 0.2,
  cor_xy = 0.5,
  xr = 0,
  yr = 0
)
}
\arguments{
\item{n}{Numeric value the number of lines of data to generate. Defaults to
500.}

\item{nwaves}{Numeric value specifying the number of waves to generate.
Defaults to 10.}

\item{ri_x}{Numeric value specifying the variance for the random intercept
for X. Defaults to 1.}

\item{ri_y}{Numeric value specifying the variance for the random intercept
for Y. Defaults to 1}

\item{cor_i}{Numeric value specifying the correlation between the random
intercepts. Defaults to .5.}

\item{x}{Numeric value specifying the variance for the autoregressive
component for X. Defaults to 1.}

\item{y}{Numeric value specifying the variance for the autoregressive
component for Y. Defaults to 1.}

\item{stab_x}{Numeric value specifying the stability of the autoregressive
process for X. Defaults to .5.}

\item{stab_y}{Numeric value specifying the stability of the autoregressive
process for Y. Defaults to .5.}

\item{yx}{Numeric value specifying the cross-lagged path predicting Y from X.
Defaults to .4.}

\item{xy}{Numeric value specifying the cross-lagged path predicting X from Y.
Defaults to .2.}

\item{cor_xy}{Numeric value specifying the correlation between the initial
autoregressive components of X and Y. Defaults to .5.}

\item{xr}{Numeric value specifying the variance of the "state" or measurement
error component for X. Defaults to 0.}

\item{yr}{Numeric value specifying the variance of the "state" or measurement
error component for Y. Defaults to 0.}
}
\value{
Dataframe with simulated data.
}
\description{
\code{gen_starts()} generates simulated data based on the STARTS model and its
variants.
}
