% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildModel.R
\name{lavaanStarts2}
\alias{lavaanStarts2}
\title{Runs Bivariate Starts Model in Lavaan}
\usage{
lavaanStarts2(
  waves,
  xWaves,
  yWaves,
  xIndicators = 1,
  yIndicators = 1,
  stationarity = TRUE,
  crossLag = TRUE,
  stateCor = TRUE,
  limits = TRUE
)
}
\arguments{
\item{waves}{Numeric value indicating total number of waves.}

\item{xWaves}{Vector of actual waves for X (omit if same as waves).}

\item{yWaves}{Vector of actual waves for Y (omit if same as waves).}

\item{xIndicators}{Numeric value indicatoring number of indicators for X.
Defaults to 1.}

\item{yIndicators}{Numeric value indicatoring number of indicators for Y.
Defaults to 1.}

\item{stationarity}{logical value indicating whether to impose stationarity.
Defaults to TRUE.}

\item{crossLag}{Logical value indicating whether to impose stationairty.
Defaults to TRUE.}

\item{stateCor}{Logical value indicating whether to constrain correlations
between wave-specific state variances to be equal across waves}

\item{limits}{Logical value indicating whether to limit variances and
correlations to valid values.}
}
\value{
Returns character vector of code that can be run using lavaan.
}
\description{
\code{lavaanStarts2()} Wrapper function for \code{run_starts_mplus()}. Produces lavaan
code for the standard bivariate STARTS model.
}
