% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildModel.R
\name{lavaanStartsY}
\alias{lavaanStartsY}
\title{Runs Univariate Starts Model for Y in Lavaan}
\usage{
lavaanStartsY(waves, yWaves, stationarity = TRUE, limits = TRUE)
}
\arguments{
\item{waves}{Numeric value indicating total number of waves.}

\item{yWaves}{Vector of actual waves for Y (omit if same as waves).}

\item{stationarity}{logical value indicating whether to impose stationarity.
Defaults to TRUE.}

\item{limits}{Logical value indicating whether to limit variances and
correlations to valid values.}
}
\value{
Returns character vector of code that can be run using lavaan.
}
\description{
\code{lavaanStartsY()} Wrapper function for \code{run_starts_mplus()}. Produces lavaan
code for the univariate STARTS model (for variabes named Y).
}
