% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{modelCode}
\alias{modelCode}
\title{Prints formatted version of lavaan or mplus code}
\usage{
modelCode(pcOutput)
}
\arguments{
\item{pcOutput}{An object created from running the \code{panelcoder()} command.}
}
\description{
\code{modelCode()} takes a panelcoder output object as an argument and then
prints corresponding lavaan or mplus code.
}
