% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelCodeR.R
\name{panelcoder}
\alias{panelcoder}
\title{Build and Run Lavaan and Mplus Code for Various Panel Models}
\usage{
panelcoder(
  data,
  title = "panelcoder",
  panelModel = "starts",
  predetermined = FALSE,
  program = "lavaan",
  crossLag = TRUE,
  ma = FALSE,
  clma = FALSE,
  traitCors = TRUE,
  arCors = TRUE,
  stateCors = FALSE,
  residCors = FALSE,
  residVar = FALSE,
  slope = "none",
  state = FALSE,
  limits = TRUE,
  rstarts = 5,
  stationarity = "paths",
  constrainState = TRUE,
  invariance = TRUE,
  mplusAnalysis = NULL,
  mplusOutput = NULL,
  mplusDirectory = "mplus",
  constrainCors = TRUE,
  run = TRUE,
  ...
)
}
\arguments{
\item{data}{Dataframe with appropriately named variables. Names should
include the name stem and wave, separated by "\emph{". If there are multiple
indicators, names should have an additional number (identifying indicator
number), again separated with "}". The number of indicators must be the
same for different waves, but different variables can have different
numbers of indicators. Mplus requires all variable names to be less than
8 characters, so variable names (including the wave and indicator indexes)
should be very short. A warning is provided if these exceed 8 characters.}

\item{title}{Title of analysis for mplus}

\item{panelModel}{Specific model to run. Can be "starts"(the default),
"riclpm", "clpm", "arts", "sts", "dpm", "gclm", "alt", "lcmsr", or
"measurement".}

\item{predetermined}{Logical value indicating whether to use a
"predetermined" version of a STARTS variant (see Andersen, 2022). This
option is not used for the DPM or GCLM.}

\item{program}{Program to use to run code. Can be "lavaan" (the default) or
"mplus"}

\item{crossLag}{Logical value indicating whether to include cross-lagged
paths. Defaults to \code{TRUE}.}

\item{ma}{Logical value indicating whether to include moving average
components in the GCLM. Defaults to \code{FALSE}.}

\item{clma}{Logical value indicating whether to include cross-lagged moving
average components in the GCLM. Defaults to \code{FALSE}.}

\item{traitCors}{Logical value indicating whether to include correlations
between stable-trait/random-intercept components. Defaults to TRUE.}

\item{arCors}{Logical value indicating whether to include correlations
between wave-specific AR components. Defaults to TRUE.}

\item{stateCors}{Logical value indicating whether to include correlations
between wave-specific state components. Defaults to FALSE.}

\item{residCors}{Logical value indicating whether to include correlations
between item-specific residual. Defaults to FALSE.}

\item{residVar}{Logical value indicating whether to constrain residual
variance to be equal across waves when there are multiple indicators.}

\item{slope}{String variable indicating what type of slope to specify in
models that include a slope. Can be "none," (the default)"linear, ""basis,"
or "centered."}

\item{state}{Logical value indicating whether to include a state component.
This is automatically included in some models (e.g., STARTS), but this
option allows for the inclusion of a state component in additional models
including the DPM and GCLM.}

\item{limits}{Logical value indicating whether to constrain variances and
correlations to possible values. Defaults to TRUE.}

\item{rstarts}{Numeric value indicating the number of random starts in
mplus. This is often necessary to avoid local maximum when limiting variance
to be positive, thus the default is to set it to 5 when limits are also set.
If you wish to remove this option, set \code{rstarts} to NULL.}

\item{stationarity}{Logical value indicating whether and how to impose
stationarity in the autoregressive process. Defaults to "paths" which
constrains the stability and cross-lagged paths to be equal. Can also be
set to "full" which also constrains variance in the autoregressive
process to be equal across waves or "none" which removes all constraints.}

\item{constrainState}{Logical value indicating whether to constrain state
variances to be equal across waves. Defaults to TRUE.}

\item{invariance}{Logical value indicating whether to constrain loadings for
the same item to be equal across waves.}

\item{mplusAnalysis}{Quoted text. Specify ANLYSIS command for mplus. Defaults
to "MODEL=NOCOVARIANCES;\\nCOVERAGE=.001;". If you change this, including
"MODEL=NOCOVARIANCES" is highly recommended given the way the model is
specified.}

\item{mplusOutput}{Quoted text. Specify OUTPUT command for mplus. Defaults to
"stdyx; \\n cinterval; \\n",}

\item{mplusDirectory}{Quoted text. Specify directory for mplus input and
output files. This directory must already exist before running the command.
Defaults to "mplus".}

\item{constrainCors}{logical value indicating whether to constrain
correlations between same indicator at different waves (when there are more
than one indicator). Not yet implemented}

\item{run}{Logical value indicating whether to run the model or to just
print and return code. Defaults to \code{TRUE}.}

\item{...}{Additional options passed to lavaan. Default options are
meanstructure=TRUE, missing = 'fiml', int.ov.free=TRUE, and
int.lv.free=FALSE.}
}
\value{
pcObject, which is a list that includes the parameter table used to
create the model, a list of basic information about the model, the actual
and implied stability coefficients used for plotting, the model code and
the fit object produced by lavaan or MplusAutomation.
}
\description{
\code{panelcoder()} produces lavaan and mplus code for variations of the general
Stable Trait, Autoregressive Trait, State Model. Univariate and bivariate
versions can be specified, and various reduced versions of the model can be
run. Other options (such as whether to include lagged paths) can also be
selected.
}
