% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{panelEstimates}
\alias{panelEstimates}
\title{Prints all estimates}
\usage{
panelEstimates(pcOutput)
}
\arguments{
\item{pcOutput}{An object created from running the \code{panelcoder()} command.}
}
\description{
\code{panelEstimates()} takes a panelcoder output object as an argument and then
prints all estimates from lavaan or mplus.
}
