% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parcel}
\alias{parcel}
\title{Creates parcels from items}
\usage{
parcel(data, items, waves, pattern)
}
\arguments{
\item{data}{Dataframe with just the items to be parcelled.}

\item{items}{Number of items for the scale.}

\item{waves}{Number of waves.}

\item{pattern}{Vector specifying the number of parcels and the number of
items per parcel. For example, to get three parcels with 3 items for the
first parcel, 3 items for the second, and 4 items for the third, specify
\code{pattern = c(3, 3, 4)}}
}
\description{
\code{parcel()} takes a set of items from a single scale and creates parcels based
on the user specification. The function conducts a factor analysis of items
at each wave and then calculates the average factor loading for each item.
The item with the strongest loading is assigned to the first parcel, the
item with the second strongest loading is assigned to the second parcel,
and so on until all items are assigned.
}
