% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.pcSum}
\alias{print.pcSum}
\title{Prints results from \code{panelcoder()}}
\usage{
\method{print}{pcSum}(x, ...)
}
\arguments{
\item{x}{Results from \code{panelcoder()}.}

\item{...}{Additional arguments to \code{print()}.}
}
\description{
Prints results from \code{panelcoder()}
}
