% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildMplus.R
\name{run_arts_mplus}
\alias{run_arts_mplus}
\title{Runs ARTS Model in Mplus}
\usage{
run_arts_mplus(
  data,
  waves,
  xWaves = NULL,
  yWaves = NULL,
  xIndicators = 1,
  yIndicators = 1,
  stationarity = TRUE,
  constrainCor = TRUE,
  trait = FALSE,
  state = TRUE,
  limits = TRUE,
  dir = "mplus",
  title = "arts",
  analysis = "MODEL=NOCOVARIANCES;\\nCOVERAGE=.001;",
  output = "stdyx; \\n  cinterval; \\n"
)
}
\arguments{
\item{data}{Dataframe that contains variables for analysis.}

\item{waves}{Numeric value indicating total number of waves.}

\item{xWaves}{Vector of actual waves for X (omit if same as waves).}

\item{yWaves}{Vector of actual waves for Y (omit if same as waves).}

\item{xIndicators}{Numeric value indicatoring number of indicators for X.
Defaults to 1.}

\item{yIndicators}{Numeric value indicatoring number of indicators for Y.
Defaults to 1.}

\item{stationarity}{logical value indicating whether to impose stationarity.
Defaults to TRUE.}

\item{constrainCor}{logical value indicating whether to constrain state
correlations to be the same.}

\item{limits}{Logical value indicating whether to limit variances and
correlations to valid values.}

\item{dir}{Character vector listing directory for mplus files. Defaults to
"mplus".}

\item{title}{Character vector. Defaults to "riclpm".}

\item{analysis}{Character vector listing the analysis statement for Mplus.
Separate options with semicolon and "\n". Defaults to
"MODEL=NOCOVARIANCES;\nCOVERAGE=.001;" .}

\item{output}{Character vector listing options for the output line in Mplus.
Separate options with semicolon and "\n". Defaults to
"stdyx; \n  cinterval; \n".}
}
\value{
Returns output from MplusAutomation::MplusModeler command, including
estimates and fit statistics.
}
\description{
\code{run_arts_mplus()} Wrapper function for \code{run_starts_mplus()}. Produces and
runs Mplus code for the ARTS Model, which is a STARTS model where the stable
trait variance is set to 0. This is also known as a "factor CLPM".
}
