% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildDpmLavaan.R
\name{run_dpm_lavaan}
\alias{run_dpm_lavaan}
\title{Build and Run Lavaan Code for Dynamic Panel Model}
\usage{
run_dpm_lavaan(
  data,
  waves,
  XVar = TRUE,
  YVar = TRUE,
  xIndicators = 1,
  yIndicators = 1,
  crossLag = TRUE,
  arCor = TRUE,
  constrainCors = TRUE,
  limits = TRUE,
  ...
)
}
\arguments{
\item{data}{Dataframe with all variables}

\item{waves}{Numeric value indicating total number of waves.}

\item{XVar}{Logical. Include X variable.}

\item{YVar}{logical. Include Y variable.}

\item{xIndicators}{Numeric value indicatoring number of indicators for X.
Defaults to 1.}

\item{yIndicators}{Numeric value indicatoring number of indicators for Y.
Defaults to 1.}

\item{crossLag}{logical value indicating whether to include cross-lagged
paths. Defaults to TRUE.}

\item{arCor}{Logical value indicating whether to include autoregresive trait
correlations. Defaults to TRUE.}

\item{constrainCors}{logical value indicating whether to constrain
correlations between same indicator at different waves (when there are more
than one indicator.}

\item{limits}{Logical value indicating whether to constrain variances and
correlations to possible values.}

\item{...}{Additional arguments to the \code{lavaan()} command.}
}
\value{
Returns pcldObject, which includes the model statement, the lavaan
object, plus extracted data for printing useful summaries.
}
\description{
\code{buildLavaanDpm()} produces lavaan code for variations of the Dynamic Panel
Model. Univariate and bivariate versions can be specified.
}
