% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildDpm.R
\name{run_dpm_mplus}
\alias{run_dpm_mplus}
\title{Runs Mplus Models}
\usage{
run_dpm_mplus(
  data,
  waves,
  XVar = TRUE,
  YVar = TRUE,
  xIndicators = 1,
  yIndicators = 1,
  crossLag = TRUE,
  constrainCors = TRUE,
  limits = TRUE,
  dir = "mplus",
  title = "dpm",
  analysis = "MODEL=NOCOVARIANCES;\\nCOVERAGE=.001;",
  output = "stdyx; \\n  cinterval; \\n"
)
}
\arguments{
\item{data}{Dataframe that contains variables for analysis.}

\item{waves}{Numeric value indicating total number of waves.}

\item{XVar}{Logical. Include X variable.}

\item{YVar}{logical. Include Y variable.}

\item{xIndicators}{Numeric value indicatoring number of indicators for X.
Defaults to 1.}

\item{yIndicators}{Numeric value indicatoring number of indicators for Y.
Defaults to 1.}

\item{crossLag}{logical value indicating whether to include cross-lagged
paths. Defaults to TRUE.}

\item{constrainCors}{logical value indicating whether to constrain
correlations between same indicator at different waves (when there are more
than one indicator.}

\item{limits}{Logical value indicating whether to limit variances and
correlations to valid values.}

\item{dir}{Character vector listing directory for mplus files. Defaults to
`.#buildMplus.R.}

\item{title}{Character vector. Defaults to \code{riclpm}.}

\item{analysis}{Character vector listing the analysis statement for Mplus.
Separate options with semicolon and \verb{\\n}. Defaults to
\verb{MODEL=NOCOVARIANCES;\\nCOVERAGE=.001;} .}

\item{output}{Character vector listing options for the output line in Mplus.
Separate options with semicolon and \verb{\\n}. Defaults to
\verb{stdyx; \\n  cinterval; \\n}.}
}
\value{
Returns output from MplusAutomation::MplusModeler command, including
estimates and fit statistics.
}
\description{
\code{run_starts_mplus()} produces code for variations of the general Stable Trait,
Autoregressive Trait, State Model and then runs the model using
MplusAutomation. Univariate and bivariate versions can be specified.
Reduced versions of the model can be specified by setting variance components
for trait, AR, or state to be \code{FALSE}.
}
