% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildModel.R
\name{run_riclpm_lavaan}
\alias{run_riclpm_lavaan}
\title{Build and Run Lavaan Code for RI-CLPM Model}
\usage{
run_riclpm_lavaan(
  data,
  waves,
  xWaves = NULL,
  yWaves = NULL,
  xIndicators = 1,
  yIndicators = 1,
  stationarity = TRUE,
  constrainCors = TRUE,
  limits = TRUE,
  ...
)
}
\arguments{
\item{data}{Dataframe with all variables.}

\item{waves}{Numeric value indicating total number of waves.}

\item{xWaves}{Vector of actual waves for X (omit if same as waves).}

\item{yWaves}{Vector of actual waves for Y (omit if same as waves).}

\item{xIndicators}{Numeric value indicatoring number of indicators for X.
Defaults to 1.}

\item{yIndicators}{Numeric value indicatoring number of indicators for Y.
Defaults to 1.}

\item{stationarity}{logical value indicating whether to impose stationarity.
Defaults to TRUE.}

\item{constrainCors}{logical value indicating whether to constrain
correlations between same indicator at different waves (when there are more
than one indicator.}

\item{limits}{Logical value indicating whether to constrain variances and
correlations to possible values.}

\item{...}{Additional arguments to the \code{lavaan()} command.}
}
\value{
Character vector representing the lavaan code for the model statement.
}
\description{
\code{run_riclpm_lavaan()} produces and runs lavaan code for the random-intercept
cross-lagged panel model.
}
