% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildDpmLavaan.R
\name{summary.pcldObject}
\alias{summary.pcldObject}
\title{Summarizes results from \code{run_dpm_lavaan()}}
\usage{
\method{summary}{pcldObject}(object, ...)
}
\arguments{
\item{object}{Results from \code{run_dpm_lavaan()}.}

\item{...}{Additional arguments to \code{summary()}.}
}
\description{
Summarizes results from \code{run_dpm_lavaan()}
}
