% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildModel.R
\name{summary.pclObject}
\alias{summary.pclObject}
\title{Summarizes results from \code{run_starts_lavaan()}}
\usage{
\method{summary}{pclObject}(object, ...)
}
\arguments{
\item{object}{Results from \code{run_starts_lavaan()}.}

\item{...}{Additional arguments to \code{summary()}.}
}
\description{
Summarizes results from \code{run_starts_lavaan()}
}
