% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildDpm.R
\name{summary.pcmdObject}
\alias{summary.pcmdObject}
\title{Summarizes results from \code{run_dpm_mplus()}}
\usage{
\method{summary}{pcmdObject}(object, ...)
}
\arguments{
\item{object}{Results from \code{run_dpm_mplus()}.}

\item{...}{Additional arguments to \code{summary()}.}
}
\description{
Summarizes results from \code{run_dpm_mplus()}
}
