% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.pcSum}
\alias{summary.pcSum}
\title{Summarizes results from \code{panelcoder()}}
\usage{
\method{summary}{pcSum}(object, ...)
}
\arguments{
\item{object}{Results from \code{panelcoder()}.}

\item{...}{Additional arguments to \code{summary()}.}
}
\description{
Summarizes results from \code{panelcoder()}
}
