% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareModels.R
\name{compareUnivariate}
\alias{compareUnivariate}
\title{Compares Univariate Variations of Models Estimated Using panelCodeR}
\usage{
compareUnivariate(
  data,
  program = "mplus",
  models = list("starts", "riclpm", "arts", "sts", "clpm"),
  title = NULL
)
}
\arguments{
\item{data}{Dataframe with multiwave data.}

\item{program}{Name of the program. Defaults to mplus as it is much quicker.}

\item{models}{List of models to compare. Defaults to list("starts", "riclpm",
"arts", "sts", "clpm"). Can also include "dpm_c", "dpm_p", "gclm", "alt",
"lgm_sm", or "lgcm" (see panelCodeR documentation).}

\item{title}{Optional character string to be attended to title of model.}
}
\description{
\code{compareUnivariate()} produces code for, runs, and compares various models for
analyzing panel data. It then uses BIC to identify the best model.
}
