% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareModels.R
\name{compareUnivariate}
\alias{compareUnivariate}
\title{Compares Univariate Variations of STARTS Model}
\usage{
compareUnivariate(data, program = "mplus", title = NULL)
}
\arguments{
\item{data}{Dataframe with multiwave data. Variables should be names 'x1' to
'xw', where 'w' is the number of waves.}

\item{program}{Name of the program. Defaults to mplus as it is much quicker.}

\item{title}{Optional character string to be attended to title of model.}
}
\description{
\code{compareUnivariate()} produces code, runs, and compares variations of the
Stable Trait, Autoregressive Trait, State Model. Specifically, it
sequentially drops each component (stable trait, autoregressive trait, and
state) and then uses BIC to identify the best model.
}
