% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getLags}
\alias{getLags}
\title{Print estimates of lagged effects}
\usage{
getLags(pcOutput)
}
\arguments{
\item{pcOutput}{output from a panelcoder run.}
}
\description{
\code{getLags()} extracts the estimates of laged effects from panelcoder output
}
