% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildModel.R
\name{lavaanClpm}
\alias{lavaanClpm}
\title{Create Code for CLPM Model in Lavaan}
\usage{
lavaanClpm(
  waves,
  xWaves = NULL,
  yWaves = NULL,
  stationarity = TRUE,
  limits = TRUE
)
}
\arguments{
\item{waves}{Numeric value indicating total number of waves.}

\item{xWaves}{Vector of actual waves for X (omit if same as waves).}

\item{yWaves}{Vector of actual waves for Y (omit if same as waves).}

\item{stationarity}{logical value indicating whether to impose stationarity.
Defaults to TRUE.}

\item{limits}{Logical value indicating whether to limit variances and
correlations to valid values.}
}
\value{
Returns character vector of code that can be run using lavaan.
}
\description{
\code{lavaanjclpm()} Wrapper function for \code{run_starts_mplus()}. Produces lavaan
code for the standard cross-lagged panel model. This is equivalent to the
STARTS model with no stable trait or state variance.
}
