% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildModel.R
\name{lavaanRiclpm}
\alias{lavaanRiclpm}
\title{Builds RI-CLPM Model in Lavaan}
\usage{
lavaanRiclpm(
  waves,
  xWaves = NULL,
  yWaves = NULL,
  xIndicators = xIndicators,
  yIndicators = yIndicators,
  stationarity = TRUE,
  constrainCors = TRUE,
  limits = TRUE
)
}
\arguments{
\item{waves}{Numeric value indicating total number of waves.}

\item{xWaves}{Vector of actual waves for X (omit if same as waves).}

\item{yWaves}{Vector of actual waves for Y (omit if same as waves).}

\item{xIndicators}{Numeric value indicatoring number of indicators for X.
Defaults to 1.}

\item{yIndicators}{Numeric value indicatoring number of indicators for Y.
Defaults to 1.}

\item{stationarity}{logical value indicating whether to impose stationarity.
Defaults to TRUE.}

\item{constrainCors}{logical value indicating whether to constrain
correlations between same indicator at different waves (when there are more
than one indicator.}

\item{limits}{Logical value indicating whether to limit variances and
correlations to valid values.}
}
\value{
Returns character vector of code that can be run using lavaan.
}
\description{
\code{lavaanRiclpm()} Wrapper function for \code{run_starts_mplus()}. Produces lavaan
code for the standard Random-Intercept Cross-Lagged Panel Model. This is
equivalent to a STARTS model without the state variance.
}
