% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelCodeR-package.R
\docType{package}
\name{panelCodeR-package}
\alias{panelCodeR}
\alias{panelCodeR-package}
\title{panelCodeR: An R Package to Create and Run Models for Analyzing Panel Data}
\description{
This package produces and then runs lavaan and Mplus code for various models for analyzing panel data.
}
\author{
\strong{Maintainer}: Richard Lucas \email{richard.e.lucas@gmail.com} (\href{https://orcid.org/0000-0002-7995-3319}{ORCID})

}
\keyword{internal}
