% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewrite.R
\name{panelcoder}
\alias{panelcoder}
\title{Build and Run Lavaan and Mplus Code for Various Panel Models}
\usage{
panelcoder(
  data,
  title = "panelcoder",
  panelModel = "starts",
  program = "lavaan",
  crossLag = TRUE,
  stateCors = FALSE,
  residCors = FALSE,
  limits = TRUE,
  stationarity = TRUE,
  invariance = TRUE,
  mplusAnalysis = NULL,
  mplusOutput = NULL,
  mplusDirectory = "mplus",
  constrainCors = TRUE,
  run = TRUE,
  ...
)
}
\arguments{
\item{data}{Dataframe with appropriately named variables. Names should
include the name stem and wave, separated by "\emph{". If there are multiple
indicators, names should have an additional number (identifying indicator
number), again separated with "}". The number of indicators must be the
same for different waves, but different variables can have different
numbers of indicators. Mplus requires all variable names to be less than
8 characters, so variable names (including the wave and indicator indexes)
should be very short. A warning is provided if these exceed 8 characters.}

\item{title}{Title of analysis for mplus}

\item{panelModel}{Specific model to run. Can be "starts"(the default),
"riclpm", "clpm", "arts", or "sts".}

\item{program}{Program to use to run code. Can be "lavaan" (the default) or
"mplus"}

\item{crossLag}{Logical value indicating whether to include cross-lagged
paths. Defaults to \code{TRUE}.}

\item{stateCors}{Logical value indicating whether to include correlations
between wave-specific state components. Defaults to FALSE.}

\item{residCors}{Logical value indicating whether to include correlations
between item-specific residual. Defaults to FALSE.}

\item{limits}{Logical value indicating whether to constrain variances and
correlations to possible values. Defaults to TRUE.}

\item{stationarity}{Logical value indicating whether to impose stationarity.
Defaults to TRUE.}

\item{invariance}{Logical value indicating whether to constrain loadings for
the same item to be equal across waves.}

\item{mplusAnalysis}{Quoted text. Specify ANLYSIS command for mplus. Defaults
to "MODEL=NOCOVARIANCES;"}

\item{mplusOutput}{Quoted text. Specify OUTPUT command for mplus. Defaults to
"stdyx; \\n cinterval; \\n",}

\item{mplusDirectory}{Quoted text. Specify directory for mplus input and
output files. This directory must already exist before running the command.
Defaults to "mplus".}

\item{constrainCors}{logical value indicating whether to constrain
correlations between same indicator at different waves (when there are more
than one indicator). Not yet implemented}

\item{run}{Logical value indicating whether to run the model or to just
print and return code.}

\item{...}{Additional options passed to lavaan. Default options are
meanstructure=TRUE and missing = 'fiml'.}
}
\value{
pcObject, which is a list that includes the parameter table used to
create the model, a list of basic information about the model, the model
code (either lavaan or mplus) and the fit object produced by lavaan or
MplusAutomation.
}
\description{
\code{panelcoder()} produces lavaan and mplus code for variations of the general
Stable Trait, Autoregressive Trait, State Model. Univariate and bivariate
versions can be specified, and various reduced versions of the model can be
run. Other options (such as whether to include lagged paths) can also be
selected.
}
