% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildMplus.R
\name{run_startsy_mplus}
\alias{run_startsy_mplus}
\title{Runs Univariate Starts Model for X in Mplus}
\usage{
run_startsy_mplus(
  data,
  waves,
  yWaves = NULL,
  yIndicators = 1,
  stationarity = TRUE,
  limits = TRUE,
  dir = "mplus",
  title = "startsy",
  analysis = "MODEL=NOCOVARIANCES;\\nCOVERAGE=.001;",
  output = "stdyx; \\n  cinterval; \\n"
)
}
\arguments{
\item{data}{Dataframe that contains variables for analysis.}

\item{waves}{Numeric value indicating total number of waves.}

\item{stationarity}{logical value indicating whether to impose stationarity.
Defaults to TRUE.}

\item{limits}{Logical value indicating whether to limit variances and
correlations to valid values.}

\item{dir}{Character vector listing directory for mplus files. Defaults to
\code{mplus}.}

\item{title}{Character vector. Defaults to \code{sts}.}

\item{analysis}{Character vector listing the analysis statement for Mplus.
Separate options with semicolon and \verb{\\n}. Defaults to
\verb{MODEL=NOCOVARIANCES;\\nCOVERAGE=.001;} .}

\item{output}{Character vector listing options for the output line in Mplus.
Separate options with semicolon and \verb{\\n}. Defaults to
\verb{stdyx; \\n  cinterval; \\n}.}

\item{xWaves}{Vector of actual waves for X (omit if same as waves).}

\item{xIndicators}{Numeric value indicatoring number of indicators for X.
Defaults to 1.}
}
\value{
Returns output from MplusAutomation::MplusModeler command, including
estimates and fit statistics.
}
\description{
\code{run_sts_mplus()} Wrapper function for \code{run_starts_mplus()}. Produces and
runs Mplus code for the STS Model, which is a STARTS model where the
autoregressive trait variance is set to 0. This is also known as a factor
CLPM.
}
