% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildMplus.R
\name{summary.pcmObject}
\alias{summary.pcmObject}
\title{Summarizes results from \code{run_starts_mplus()}}
\usage{
\method{summary}{pcmObject}(obj, ...)
}
\arguments{
\item{obj}{Results from \code{run_starts_mplus()}.}
}
\description{
Summarizes results from \code{run_starts_mplus()}
}
